refactor: Constants 클래스 utils 패키지에서 이전
