+++ src/main/java/racingcar/Car.java
@@ -1,6 +1,5 @@
 package racingcar;
 
-import utils.ConstantUtils;
 import utils.RandomUtils;
 
 public class Car {
@@ -14,8 +13,8 @@ public class Car {
 
     public void moveOrNot() {
         int computerNumber = RandomUtils
-            .nextInt(ConstantUtils.MIN_RANDOM_ZERO, ConstantUtils.MAX_RANDOM_NINE);
-        if (computerNumber > ConstantUtils.MAX_NO_MOVE) {
+            .nextInt(Constants.MIN_RANDOM_ZERO, Constants.MAX_RANDOM_NINE);
+        if (computerNumber > Constants.MAX_NO_MOVE) {
             position += 1;
         }
     }
