+++ src/main/java/racingcar/PlayerResponse.java
@@ -1,7 +1,6 @@
 package racingcar;
 
 import java.util.Scanner;
-import utils.ConstantUtils;
 
 public class PlayerResponse {
 
@@ -17,7 +16,7 @@ public class PlayerResponse {
         String[] carNames;
         while (true) {
             try {
-                System.out.println(ConstantUtils.ASKING_NAMES_MESSAGE);
+                System.out.println(Constants.ASKING_NAMES_MESSAGE);
                 carNames = responseOfCarNames();
                 validator.checkCarNames(carNames);
                 return carNames;
@@ -30,7 +29,7 @@ public class PlayerResponse {
     public int getHowManyMoves() {
         while (true) {
             try {
-                System.out.println(ConstantUtils.ASKING_MOVES_MESSAGE);
+                System.out.println(Constants.ASKING_MOVES_MESSAGE);
                 String movesBeforeCheck = responseOfMoves();
                 validator.checkNumber(movesBeforeCheck);
                 return Integer.parseInt(movesBeforeCheck);
@@ -41,7 +40,7 @@ public class PlayerResponse {
     }
 
     private String[] responseOfCarNames() {
-        String[] carNames = scanner.nextLine().split(ConstantUtils.NAMES_DELIMITER, -1);
+        String[] carNames = scanner.nextLine().split(Constants.NAMES_DELIMITER, -1);
         return deleteWhiteSpace(carNames);
     }
