+++ src/main/java/racingcar/RacingCarGame.java
@@ -2,7 +2,6 @@ package racingcar;
 
 import java.util.ArrayList;
 import java.util.List;
-import utils.ConstantUtils;
 
 public class RacingCarGame {
 
@@ -29,7 +28,7 @@ public class RacingCarGame {
 
     private void startRacing() {
         System.out.println();
-        System.out.println(ConstantUtils.RACE_RESULT);
+        System.out.println(Constants.RACE_RESULT);
         for (int i = 0; i < moves; i++) {
             oneLap();
             System.out.println();
@@ -45,14 +44,14 @@ public class RacingCarGame {
     }
 
     private void printResultOfEachCar(Car car) {
-        System.out.print(car.getName() + ConstantUtils.NAME_RESULT_SEPARATOR);
+        System.out.print(car.getName() + Constants.NAME_RESULT_SEPARATOR);
         printPosition(car);
         System.out.println();
     }
 
     private void printPosition(Car car) {
         for (int j = 0; j < car.getPosition(); j++) {
-            System.out.print(ConstantUtils.RACING_BAR);
+            System.out.print(Constants.RACING_BAR);
         }
     }
 
@@ -63,9 +62,9 @@ public class RacingCarGame {
     }
 
     private void announceWinner() {
-        System.out.print(ConstantUtils.FINAL_WINNER);
+        System.out.print(Constants.FINAL_WINNER);
         addWinnerList();
-        System.out.println(String.join(ConstantUtils.WINNERS_JOIN_DELIMITER, winnerList));
+        System.out.println(String.join(Constants.WINNERS_JOIN_DELIMITER, winnerList));
     }
 
     private void addWinnerList() {
