+++ src/main/java/racingcar/Validator.java
@@ -1,7 +1,5 @@
 package racingcar;
 
-import utils.ConstantUtils;
-
 public class Validator {
 
     public Validator() {
@@ -10,11 +8,11 @@ public class Validator {
 
     public void checkCarNames(String[] carNames) {
         for (String carName : carNames) {
-            if (carName.length() > ConstantUtils.MAX_CAR_NAME) {
-                throw new IllegalArgumentException(ConstantUtils.OVER_FIVE_LETTERS);
+            if (carName.length() > Constants.MAX_CAR_NAME) {
+                throw new IllegalArgumentException(Constants.OVER_FIVE_LETTERS);
             }
-            if (carName.length() < ConstantUtils.MIN_CAR_NAME) {
-                throw new IllegalArgumentException(ConstantUtils.EMPTY_LETTER);
+            if (carName.length() < Constants.MIN_CAR_NAME) {
+                throw new IllegalArgumentException(Constants.EMPTY_LETTER);
             }
         }
     }
@@ -22,11 +20,11 @@ public class Validator {
     public void checkNumber(String num) {
         try {
             int number = Integer.parseInt(num);
-            if (number < ConstantUtils.MIN_LAP) {
-                throw new IllegalArgumentException(ConstantUtils.MUST_BE_POSITIVE_INTEGER);
+            if (number < Constants.MIN_LAP) {
+                throw new IllegalArgumentException(Constants.MUST_BE_POSITIVE_INTEGER);
             }
         } catch (NumberFormatException n) {
-            throw new IllegalArgumentException(ConstantUtils.MUST_BE_NUMBER);
+            throw new IllegalArgumentException(Constants.MUST_BE_NUMBER);
         }
     }
 }
